move ancestry main code to main.py this should resolve circular import issue
